refactor all document stuff in the same module
