+++ mindee/documents/document_config.py
@@ -1,8 +1,10 @@
+from typing import Dict
+from mindee.documents.base import OFF_THE_SHELF, CUSTOM_DOCUMENT
 from mindee.documents.custom_document import CustomDocument
 
 
 class DocumentConfig:
-    def __init__(self, config: dict, doc_type: str = "custom_document"):
+    def __init__(self, config: dict, doc_type: str = CUSTOM_DOCUMENT):
         """
         :param config: (dict) Object containing config
         :param doc_type: (string) off_the_shelf or custom_document
@@ -19,7 +21,7 @@ class DocumentConfig:
         self.singular_name = config["singular_name"]
         self.plural_name = config["plural_name"]
 
-        if self.type == "custom_document":
+        if self.type == CUSTOM_DOCUMENT:
             # Check for endpoint URL and API Key in config
             for mandatory_field in ["endpoint", "api_key"]:
                 if mandatory_field not in config.keys():
@@ -28,25 +30,28 @@ class DocumentConfig:
                         % mandatory_field
                     )
             self.constructor = CustomDocument
-            self.api_key_kwargs = None
+            self.api_key_kwargs = []
             self.api_key = config["api_key"]
             self.endpoint = config["endpoint"]
-        elif self.type == "off_the_shelf":
+        elif self.type == OFF_THE_SHELF:
             self.constructor = config["constructor"]
             self.api_key_kwargs = config["api_key_kwargs"]
 
-    @staticmethod
-    def validate_list(config_list):
-        """Validate the configuration list definitions."""
-        if len(set([v.singular_name for k, v in config_list.items()])) != len(
-            [v.singular_name for k, v in config_list.items()]
-        ):
-            raise AssertionError(
-                "singular_name values must be uniques among custom_documents list objects"
-            )
-        if len(set([v.plural_name for k, v in config_list.items()])) != len(
-            [v.plural_name for k, v in config_list.items()]
-        ):
-            raise AssertionError(
-                "plural_name values must be uniques among custom_documents list objects"
-            )
+
+DocumentConfigDict = Dict[str, DocumentConfig]
+
+
+def validate_list(config_list: DocumentConfigDict):
+    """Validate the configuration list definitions."""
+    if len(set([v.singular_name for k, v in config_list.items()])) != len(
+        [v.singular_name for k, v in config_list.items()]
+    ):
+        raise AssertionError(
+            "singular_name values must be uniques among custom_documents list objects"
+        )
+    if len(set([v.plural_name for k, v in config_list.items()])) != len(
+        [v.plural_name for k, v in config_list.items()]
+    ):
+        raise AssertionError(
+            "plural_name values must be uniques among custom_documents list objects"
+        )
