+++ mindee/fields/date.py
@@ -16,7 +16,7 @@ class Date(Field):
         :param reconstructed: Bool for reconstructed object (not extracted in the API)
         :param page_n: Page number for multi pages pdf
         """
-        super(Date, self).__init__(
+        super().__init__(
             date_prediction,
             value_key=value_key,
             reconstructed=reconstructed,
