+++ mindee/fields/orientation.py
@@ -15,7 +15,7 @@ class Orientation(Field):
         :param reconstructed: Bool for reconstructed object (not extracted in the API)
         :param page_n: Page number for multi pages pdf
         """
-        super(Orientation, self).__init__(
+        super().__init__(
             orientation_prediction,
             value_key=value_key,
             reconstructed=reconstructed,
