+++ tests/documents/test_document.py
@@ -2,7 +2,7 @@ import pytest
 import json
 from mindee import Response, Client
 from mindee.inputs import Inputs
-from mindee.documents import Document
+from mindee.documents.base import Document
 
 
 @pytest.fixture
